Cleaned up output handling of urls
All output is now handled by the driver function and not passed on
to the called function.  They, instead, return flattened lists of
URLs that are concatenated and written to the output file after
all parsing is complete.  This resulted in reduced code duplication
and less clutter in the parsing functions.