Added lots of extra functionality/padding/error responses for the Elsevier Fulltext notebook.
Vast majority of the additions in this commit had to do with the for loop which loops through the URL list and actually grabs the papers. I completely restructured this loop so that if for some reason retrieval of a single type of metadata from the elsapy object doesn't work, other pieces of data from that same object can still be collected.

The way I accomplished this was by splitting up the retrieval of those pieces of data into various try/except clauses.