Many changes to the Science Direct URL scraper. Vastly increased functionality.
The URL scraper mechanism of search has been altered.
The new mechanism is to find out what the current list of all journals that sciencedirect owns.
Once you know the titles, pick out the journals that you think are relewvant for the publications you are looking for.
This is done by keyword searching the journal titles for your subject.
In the current version of the scraper, it looks for key substring matches like 'biochem' and 'energy'

Once those journals are known, the scraper creates a triple-nested dictionary to construct a url for
every possible permutation of (1) a keyword search (2) a year (3) journal name (4) page number.