Added two scripts to turn metacorpus into fulltext corpus.
Elsevier_fulltext_api.py acts on a metadata corpus built by 'pybliometrics.ipynb' and creates a fulltext corpus.
run_fulltext.sh is a bash script that parallelizes the fulltext builder mentioned above.

This version of the fulltext corpus builder was able to build a corpus of ~800K-1M full texts (haven't counted in a while).