Added lots of NER processing scripts/notebooks
At this point, ner_sheets.py has the ability to collect all labeled NER data
from A SINGLE EXCEL SHEET into a pandas dataframe.

Now the goal is to make a function that does this for every single sheet in a folder,
recursively, and stacks all the training data into an array of tuples.