ner.py now makes training_data from all folders!
This is the first successful version of ner label extraction that loops through all NER folders
and consolodates into one training data dictionary.

Next step is to make sure the arch_blocks are good, and probably
create a module called training.py that allows a user to abstract
training commands.