committing ner.py before making w2v additions
The w2v additions being referenced are:
Having the NerData class take in a word2vec model to build a vocabulary with.
This is needed so that words with good embeddings that ARE NOT in the training set
can still be sent through the lstm model.

Also contained in this commit:
minor changes to try and make a 'convertables' dictionary
so that every sentence in the training data is read as a string
by pd.read_excel. This was an attempt to solve problem of a token
'9-19' in the training data being read as a datetime.datetime object,
rather than a string.